Changed city to location in FilteredCity
